convert test_prevector.cpp
