+++ src/Makefile.test.include
@@ -46,7 +46,6 @@ BITCOIN_TESTS =\
   test/checkblock_tests.cpp \
   test/coins_tests.cpp \
   test/dbwrapper_tests.cpp \
-  test/prevector_tests.cpp \
   test/reverselock_tests.cpp \
   test/rpc_tests.cpp \
   test/scheduler_tests.cpp \
